+++ CHANGELOG.md
@@ -5,6 +5,7 @@ HEAD
 ----
 
 * Fixed warning "unused variable" with GCC 4.4 (issue #912)
+* Fixed warning "maybe uninitialized" (issue #909)
 
 v5.13.4
 -------