+++ CHANGELOG.md
@@ -4,6 +4,7 @@ ArduinoJson: change log
 HEAD
 ----
 
+* Fixed warning "unused variable" with GCC 4.4 (issue #912)
 * Added a clear error message for `StaticJsonBuffer` and `DynamicJsonBuffer`
 
 v6.9.0 (2019-02-26)